inv/requirements.txt
Flask==3.1.1
pyOpenSSL==25.1.0
watchdog==6.0.0